convert sting into JSX
